Play optimal moves even when loss/win is certain. Major refactor.

+++ search.h
@@ -10,9 +10,9 @@ using namespace std;
 constexpr bitboard VALID_MOVE_MASK = 279258638311359;
 
 typedef struct result {
-  int move;
+  short move;
   int value;
-  int depth;
+  short depth;
 } result_t;
 
 result_t miniMax(Board b, int depth);
@@ -20,6 +20,6 @@ result_t getBestMove(Board b, int depth);
 std::vector<result_t> getPossibleResults(Board b, int depth);
 int getValue(bitboard* bitboards, bool previous_player);
 int countWins(bitboard board);
-result_t makeResult(int move, int value, int depth);
+result_t makeResult(short move, int value, short depth);
 
 #endif
